% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\docType{methods}
\name{as.character,pathlist-method}
\alias{as.character,pathlist-method}
\title{Character Representation of a pathlist Object}
\usage{
\S4method{as.character}{pathlist}(x, relative = FALSE)
}
\arguments{
\item{x}{a pathlist object}

\item{relative}{if \code{TRUE} (the default is \code{FALSE}) the root is
omitted from the paths}
}
\description{
Character Representation of a pathlist Object
}
