% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{as.list,pathlist-method}
\alias{as.list,pathlist-method}
\title{List Representation of a pathlist Object}
\usage{
\S4method{as.list}{pathlist}(x, relative = FALSE)
}
\arguments{
\item{x}{a pathlist object}

\item{relative}{if \code{TRUE} (the default is \code{FALSE}) the root is
omitted from the paths}
}
\description{
List Representation of a pathlist Object
}
