% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\docType{methods}
\name{as.matrix,pathlist-method}
\alias{as.matrix,pathlist-method}
\title{Matrix Representation of a pathlist Object}
\usage{
\S4method{as.matrix}{pathlist}(x, relative = FALSE)
}
\arguments{
\item{x}{a pathlist object}

\item{relative}{if \code{TRUE} (the default is \code{FALSE}) the root is
removed from the paths}
}
\description{
Matrix Representation of a pathlist Object
}
