% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\docType{methods}
\name{$,pathlist-method}
\alias{$,pathlist-method}
\title{Filter Paths by Top Level Folder}
\usage{
\S4method{$}{pathlist}(x, name)
}
\arguments{
\item{x}{object of class pathlist}

\item{name}{name of top level folder}
}
\description{
Filter Paths by Top Level Folder
}
