% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\docType{methods}
\name{depth,pathlist-method}
\alias{depth,pathlist-method}
\title{Get the File Path Depths}
\usage{
\S4method{depth}{pathlist}(object)
}
\arguments{
\item{object}{object of class pathlist}
}
\description{
Get the File Path Depths
}
