% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{.DollarNames.pathlist}
\alias{.DollarNames.pathlist}
\title{Provide Names for Autocomplete after Dollar}
\usage{
\method{.DollarNames}{pathlist}(x, pattern)
}
\arguments{
\item{x}{object}

\item{pattern}{patterm matching possible completions}
}
\description{
Provide Names for Autocomplete after Dollar
}
