% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{filename,pathlist-method}
\alias{filename,pathlist-method}
\title{Get the File Name Only}
\usage{
\S4method{filename}{pathlist}(object)
}
\arguments{
\item{object}{object of class pathlist}
}
\description{
Get the File Name Only
}
