% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{folder,pathlist-method}
\alias{folder,pathlist-method}
\title{Get the Folder Path of all Folders Below the Top-Level}
\usage{
\S4method{folder}{pathlist}(object)
}
\arguments{
\item{object}{object of class pathlist}
}
\description{
Get the Folder Path of all Folders Below the Top-Level
}
