% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guess_and_set_declared_encodings.R
\name{guess_and_set_declared_encodings}
\alias{guess_and_set_declared_encodings}
\title{Guess and Set the Declared Encodings}
\usage{
guess_and_set_declared_encodings(x)
}
\arguments{
\item{x}{vector of character}
}
\description{
Guess and Set the Declared Encodings
}
\examples{
Encoding("\xe4")
Encoding(guess_and_set_declared_encodings("\xe4"))

}
