% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\docType{methods}
\name{head,pathlist-method}
\alias{head,pathlist-method}
\title{First Paths in pathlist}
\usage{
\S4method{head}{pathlist}(x, n = 6)
}
\arguments{
\item{x}{object of class pathlist}

\item{n}{number of paths}
}
\description{
First Paths in pathlist
}
