% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{hide_server,pathlist-method}
\alias{hide_server,pathlist-method}
\title{Hide the Server Name in the Root}
\usage{
\S4method{hide_server}{pathlist}(object)
}
\arguments{
\item{object}{object of class pathlist}
}
\description{
Hide the Server Name in the Root
}
