% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{length,pathlist-method}
\alias{length,pathlist-method}
\title{Number of Paths in pathlist}
\usage{
\S4method{length}{pathlist}(x)
}
\arguments{
\item{x}{object of class pathlist}
}
\description{
Number of Paths in pathlist
}
