% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\docType{class}
\name{pathlist-class}
\alias{pathlist-class}
\alias{pathlist}
\title{Class to Work with File Paths}
\description{
This class simplifies the work with file paths. Paths are internally stored
as a matrix of subfolders with one row per file and one column per path depth
level. You can use the Dollar operator to filter for paths belonging to a
certain top folder.
}
