% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\docType{methods}
\name{show,pathlist-method}
\alias{show,pathlist-method}
\title{Print a pathlist Object}
\usage{
\S4method{show}{pathlist}(object)
}
\arguments{
\item{object}{a pathlist object}
}
\description{
Print a pathlist Object
}
