% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\docType{methods}
\name{[,pathlist-method}
\alias{[,pathlist-method}
\title{Filter Paths by Index}
\usage{
\S4method{[}{pathlist}(x, i)
}
\arguments{
\item{x}{object of class pathlist}

\item{i}{indices in the vector of paths}
}
\description{
Filter Paths by Index
}
