% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{summary,pathlist-method}
\alias{summary,pathlist-method}
\title{Print Path Summary}
\usage{
\S4method{summary}{pathlist}(object)
}
\arguments{
\item{object}{object of class pathlist}
}
\description{
This function prints a summary of a vector of file paths giving the common
root of all paths, the number of paths, the maximum path depth, the number
of paths within each depth and the names of the ten top level folders that
are biggest in terms of number of contained files
}
