% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{tail,pathlist-method}
\alias{tail,pathlist-method}
\title{Last Paths in pathlist}
\usage{
\S4method{tail}{pathlist}(x, n = 6)
}
\arguments{
\item{x}{object of class pathlist}

\item{n}{number of paths}
}
\description{
Last Paths in pathlist
}
