% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\docType{methods}
\name{toplevel,pathlist-method}
\alias{toplevel,pathlist-method}
\title{Get the Top-Level Folders}
\usage{
\S4method{toplevel}{pathlist}(object)
}
\arguments{
\item{object}{object of class pathlist}
}
\description{
Get the Top-Level Folders
}
